requirements.txt.txt
pandas==2.0.3
numpy==1.23.5
matplotlib==3.7.2
scikit-learn==1.2.2
pyserial==3.5
